+++ components/SdCardViewer.tsx
@@ -1,8 +1,5 @@
-import { Loader } from "@mantine/core";
-
 import { SdImageComp } from "./SdImageComp";
 import { SdImagePlaceHolderComp } from "./SdImagePlaceHolderComp";
-import { useInfiniteScroll } from "./useInfiniteScroll";
 
 import type { SdImage, SdImagePlaceHolder } from "../libs/shared-types/src";
 
@@ -11,16 +8,18 @@ type SdCardViewerProps = {
   placeholderImages: SdImagePlaceHolder[];
 
   childCard?: React.ReactNode;
+  childCard2?: React.ReactNode;
 };
 
 export function SdCardViewer(props: SdCardViewerProps) {
-  const { imageGroupData, childCard, placeholderImages } = props;
+  const { imageGroupData, childCard, placeholderImages, childCard2 } = props;
 
   const imageSize = 512;
 
   return (
     <div className="mb-4 grid grid-cols-1 gap-2 sm:grid-cols-3 md:grid-cols-5">
       {childCard}
+      {childCard2}
       {placeholderImages.map((item: SdImagePlaceHolder) => (
         <div key={item.id}>
           <SdImagePlaceHolderComp